museum fix & another tool
change of Parameters allows to retrieve a list of trafos, not just one.

The CompareTrafos tool takes all specified trafos, applies them to the (specified) input source, and then reports whenever not all trafos produce outputs of the same size. That helped to track down the issues in Museum. This time there was an issue with a too fancy equivalence which I simply replaced with ===.