divide-and-conquer fixed
fixed some issues with the divide-and-conquer strategy; FactSearch is a new module for testing the soundness of such strategies.  Effect changed slightly so that it just outputs the geo means, and the average only under option -v. EffectData is a program to extract effect tables from an allEffects run. This is mostly doing the same things (but on different data) than RunTimes, so most of their code is shared.