HUExtract etc
HUExtract is a new program that produces the sample set of HU-expanded REs to stdout, depending on params. MinRatios2 is the back end for this that computes the ratio and assumes that the expanded REs are stored in the pop of their original size. RegexpCount has tools to figure out what proportion of REs of a given size are (guaranteed to be) the toal lang, or for what proportion we have alpha(x)=swp(x).