RegexpCount improvement
Reg with option -r reprots now some information about the target population. In particular, how big it is and what percentage of its expression will be simplified by lifting to the total language. This info takes a bit longer to compute than the normal distribution (and it takes longer for larger alphabets too), so it's not recommended for S=2560 and W=16.